fix: fix calendar metrics
