+++ frontend/messages/en.json
@@ -373,6 +373,15 @@
     "month": "Month",
     "week": "Week",
     "timeline": "Timeline",
+    "previous": "Previous",
+    "next": "Next",
+    "allDay": "All day",
+    "day": "Day",
+    "date": "Date",
+    "time": "Time",
+    "event": "Event",
+    "noEventsInRange": "No events in range",
+    "showMore": "+{count} more",
     "filters": "Filters",
     "newRequest": "New Request",
     "proTip": "Pro-Tip"
