+++ frontend/messages/es.json
@@ -377,6 +377,15 @@
     "month": "Mes",
     "week": "Semana",
     "timeline": "Cronograma",
+    "previous": "Anterior",
+    "next": "Siguiente",
+    "allDay": "Todo el día",
+    "day": "Día",
+    "date": "Fecha",
+    "time": "Hora",
+    "event": "Evento",
+    "noEventsInRange": "No hay eventos en este rango",
+    "showMore": "+{count} más",
     "filters": "Filtros",
     "newRequest": "Nueva Solicitud",
     "proTip": "Pro-Tip"
